requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
sqlcipher3==0.5.2
alembic==1.13.1

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.2

# Web Scraping & Automation
playwright==1.41.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
httpx==0.26.0
aiohttp==3.10.10

# Social Media Tools
snscrape==0.7.0.20230622
telethon==1.34.0
yt-dlp==2024.1.7

# Facial Recognition & AI
face-recognition==1.3.0
deepface==0.0.87
opencv-python==4.9.0.80
dlib==19.24.2
Pillow==10.2.0

# NLP & Sentiment Analysis
transformers==4.37.2
torch==2.2.0
textblob==0.17.1
nltk==3.8.1
wordcloud==1.9.3

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# Report Generation
jinja2==3.1.3
weasyprint==60.2
reportlab==4.0.9
qrcode==7.4.2
pypdf==4.0.1

# Email & Phone Validation
phonenumbers==8.13.27
email-validator==2.1.0.post1

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2024.1
schedule==1.2.1
tqdm==4.66.1
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
